refactor: import UserAuthResponseDto from the correct place tm-25
