--- apps/frontend/src/pages/overview/libs/components/welcome-section-header/libs/types.ts
@@ -1,1 +1,0 @@
-export { UserAuthResponseDto } from "shared";
