+++ apps/frontend/src/pages/overview/libs/components/welcome-section-header/welcome-section-header.tsx
@@ -1,7 +1,7 @@
 import welcomeCharacter from "~/assets/img/svg/welcome-character.svg";
 import { Button } from "~/libs/components/components.js";
+import { UserAuthResponseDto } from "~/modules/users/users.js";
 
-import { UserAuthResponseDto } from "./libs/types.js";
 import styles from "./styles.module.css";
 
 type Properties = {
