--- apps/frontend/src/pages/overview/libs/types/types.ts
@@ -1,1 +1,0 @@
-export { type UserAuthResponseDto } from "shared";
