+++ apps/frontend/src/pages/overview/overview.tsx
@@ -1,7 +1,7 @@
 import { useAppSelector } from "~/libs/hooks/hooks.js";
+import { UserAuthResponseDto } from "~/modules/users/users.js";
 
 import { WelcomeHeader } from "./libs/components/components.js";
-import { UserAuthResponseDto } from "./libs/types/types.js";
 
 const Overview: React.FC = () => {
 	const { user } = useAppSelector((state) => state.auth);
